backend/lib/kre8vidmems/requirements.txt
numpy
opencv-python
qrcode[pil]
sentence-transformers
annoy
tqdm
Pillow
pyzbar
openai
google-generativeai
anthropic
